feat: add player object and build screens from there

+++ app/src/main/java/com/example/idlecorporationclicker/model/IBuilding.kt
@@ -6,13 +6,13 @@ import com.badlogic.gdx.scenes.scene2d.ui.Image
 interface IBuilding {
     abstract val type: BuildingType
     abstract val name: String
-    abstract var value: Int
-    abstract var level: Int
-    abstract var upgradeCost: Int
+    abstract var value: Double
+    abstract var level: Double
+    abstract var upgradeCost: Double
     abstract val image : Image
     abstract fun upgrade()
-    abstract fun calculateValue() : Int
-    abstract fun calculateUpgradeCost() : Int
+    abstract fun calculateValue() : Double
+    abstract fun calculateUpgradeCost() : Double
 }
 
 enum class BuildingType {
